ci: automate release process (#6)
This is a blatant copy of https://github.com/Doist/reactist/pull/863,
hence my targeted review request for Rui.

I guess I now have to configure the secrets (`GITHUB_TOKEN` and
`NPM_TOKEN`), which I'm not sure where to get from. ⚠️ **Also, please
know that this repository may eventually be public, so avoid discussing
anything sensitive here.**

Is this all it takes? Will this work out-of-the-box after that is
configured and this gets merged? Or is there something else to be done?

~~For instance, if I merge this before #5, and then I merge that PR,
will that trigger the v0.2.0 release? _(Because it's a `feat` I assume
it will increase in that way, but I'm not sure about semver on v0.x
which behaves differently; but you get my point.)_~~

I forgot that I merged #5 beforehand, so now I do not have a pull
request with which to test this. But I still need to get the changes in
#5 released. What do you think I should do? Release those changes first
manually?