feat!: Remove single update task and replace with multiple (#47)
# Pull Request

## Short description

Nearly all of our tools for updating a task (completing, organising,
etc) allow for multiple tasks to be sent through, but doing a regular
update does not. This PR removes the single task update tool and
replaces it with a tool that allows for updating multiple tasks in one
go.

## PR Checklist

Feel free to leave unchecked or remove the lines that are not
applicable.

-   [x] Added tests for bugs / new features
-   [ ] Updated docs (README, etc.)

<!--
_Note:_ versioning is handled by
[release-please](https://github.com/googleapis/release-please) action,
based on the PR title.
-->