feat: export comment tools (#63)
## Short description

In #58 I forgot to export the new comment tools in the `src/index.ts`
file. This is of little consequence, as they were included in the
`getMcpServer` convenience function, so any MCP server application that
updated to v4.0.0 would still get the new tools. But it is still
important to formally export them for individual use.

This PR also adds a new checkbox to the GitHub PR template to remind PR
authors to do this when needed. This hopefully minimizes the chances of
this happening again.

## PR Checklist

- [x] New tools added to `getMcpServer` AND exported in `src/index.ts`.