feat: Add support for filtering by label (#71)
# Pull Request

## Short description

A suggestion from @amix was that we should be able to filter task
acquisition by labels. This adds support to the three main task tools
such that labels can be provided, and it can be set whether a task
should have any of the labels or all of them.

## PR Checklist

Feel free to leave unchecked or remove the lines that are not
applicable.

-   [ ] Added tests for bugs / new features
-   [ ] Updated docs (README, etc.)
- [ ] New tools added to `getMcpServer` AND exported in `src/index.ts`.

<!--
_Note:_ versioning is handled by
[release-please](https://github.com/googleapis/release-please) action,
based on the PR title.
-->